rename module from _cached to _sql
