+++ airbyte-lib/airbyte_lib/datasets/__init__.py
@@ -1,7 +1,7 @@
 from airbyte_lib.datasets._base import DatasetBase
-from airbyte_lib.datasets._cached import CachedDataset, SQLDataset
 from airbyte_lib.datasets._lazy import LazyDataset
 from airbyte_lib.datasets._map import DatasetMap
+from airbyte_lib.datasets._sql import CachedDataset, SQLDataset
 
 
 __all__ = [
